# internal/infrastructure/http/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stori Challenge API
  version: "1.0.0"
  description: "API for CSV migration and processing. Base URL prefix: /v1"
servers:
  - url: /v1
paths:
  /migrate:
    post:
      summary: Migrate transactions via CSV upload
      description: "Accepts a CSV file with columns id,user_id,amount,datetime and migrates transactions. Endpoint: POST /v1/migrate"
      tags:
        - migrate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  value:
                    inserted: 120
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCsv:
                  value:
                    code: invalid_csv
                    message: CSV validation failed
                    errors:
                      - row: 3
                        field: amount
                        value: "abc"
                        message: must be a number
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conflict:
                  value:
                    code: conflict
                    message: Conflicting transactions
                    errors:
                      - row: 15
                        field: id
                        value: "tx-123"
                        message: duplicate transaction id
  /users/{user_id}/balance:
    get:
      summary: Get user balance summary within an optional time range
      description: "Returns balance, total_debits and total_credits for a user. Query params must be RFC3339 with Z."
      tags:
        - users
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: User ID
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
          description: "Lower bound (RFC3339 with Z). If only one bound is provided, this is used as lower bound and now is used as upper bound."
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
          description: "Upper bound (RFC3339 with Z). If both are provided, bounds are ordered automatically."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
              examples:
                ok:
                  value:
                    balance: 25.21
                    total_debits: 10
                    total_credits: 15
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidDate:
                  value:
                    code: invalid_datetime
                    message: datetime must be RFC3339 with Z
                    errors: []
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    code: user_transactions_not_found
                    message: user has no transactions
components:
  schemas:
    SuccessResponse:
      type: object
      properties:
        inserted:
          type: integer
          description: Number of inserted transactions
      required:
        - inserted
    ErrorItem:
      type: object
      properties:
        row:
          type: integer
        field:
          type: string
        value:
          type: string
        message:
          type: string
      required:
        - field
        - message
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable application error code
        message:
          type: string
          description: Human-readable error message
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
      required:
        - code
        - message
        - errors
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
          format: double
        total_debits:
          type: number
          format: double
        total_credits:
          type: number
          format: double
      required:
        - balance
        - total_debits
        - total_credits